Create Spreadsheets Automation
This code first imports the openpyxl library, which is used to work with Excel files. Then, it defines a function called create_pivot_table(), which takes the path to the Excel file, the name of the sheet, the range of cells that contains the data for the pivot table, and a list of the fields that should be included in the pivot table as input. The function then creates a pivot table in the specified sheet and saves the workbook.

The main() function calls the create_pivot_table() function with the appropriate arguments.

You can use this code as a starting point to automate other tasks in Excel, such as creating charts, formulas, and formatting.